Reorganised request/response model in API layer.
