+++ Warehouse.API/Controllers/ProductsController.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Warehouse.Application.Products.Queries;
 using Warehouse.Domain.Entities;
-using Warehouse.API.QueryParams;
+using Warehouse.API.Models.Requests;
 using MediatR;
 
 namespace Warehouse.API.Controllers;
@@ -14,7 +14,7 @@ public class ProductsController : BaseController
     // GET: api/<ProductController>
     [HttpGet]
     [Route("api/filter")]
-    public async Task<ActionResult<IList<Product>>> GetFiltered([FromQuery] ProductFilter filter)
+    public async Task<ActionResult<IList<Product>>> GetFiltered([FromQuery] ProductFilterRequest filter)
     {
         var getPerson = new GetProductsFiltered { MinPrice = filter.MinPrice, MaxPrice = filter.MaxPrice, Highlights = filter.Highlights};
         var result = await _mediator.Send(getPerson);
