+++ Warehouse.API/Warehouse.API.csproj
@@ -34,9 +34,15 @@
   <ItemGroup>
     <None Remove="QueryParams\" />
     <None Remove="Logs\" />
+    <None Remove="Models\" />
+    <None Remove="Models\Requests\" />
+    <None Remove="Models\Responses\" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="QueryParams\" />
     <Folder Include="Logs\" />
+    <Folder Include="Models\" />
+    <Folder Include="Models\Requests\" />
+    <Folder Include="Models\Responses\" />
   </ItemGroup>
 </Project>
